# src/swagger.yaml
openapi: 3.0.3
info:
  title: MedVehicule API
  description: API REST de gestion de flotte de véhicules hospitaliers
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/vehicles:
    get:
      summary: Liste des véhicules
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: hospitalId
          in: query
          schema: { type: string }
        - name: vehicleTypeId
          in: query
          schema: { type: string }
        - name: available
          in: query
          schema: { type: string, enum: [true, false] }
      responses:
        '200': { description: Liste }
    post:
      summary: Créer un véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [hospitalId, vehicleTypeId]
              properties:
                hospitalId: { type: string }
                vehicleTypeId: { type: string }
                registration: { type: string }
                brand: { type: string }
                model: { type: string }
                year: { type: integer }
      responses:
        '201': { description: Créé }
  /api/vehicles/{vehicleId}:
    get:
      summary: Détail véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
    patch:
      summary: Modifier véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                registration: { type: string }
                brand: { type: string }
                model: { type: string }
                year: { type: integer }
                vehicleTypeId: { type: string }
      responses:
        '200': { description: OK }
    delete:
      summary: Supprimer véhicule
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Supprimé }
  /api/vehicles/{vehicleId}/history:
    get:
      summary: Historique du véhicule
      description: Prises, interventions et entretiens
      tags: [Véhicules]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/photos/vehicle/{vehicleId}:
    get:
      summary: Photos d'un véhicule
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
    post:
      summary: Ajouter une photo
      description: multipart/form-data - photo (fichier), label, bookingId
      tags: [Photos]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo: { type: string, format: binary }
                label: { type: string }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
...
  /api/odometer/vehicle/{vehicleId}:
    get:
      summary: Relevés compteur
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200': { description: Liste }
    post:
      summary: Enregistrer un relevé
      tags: [Compteur]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value: { type: integer, minimum: 0 }
                bookingId: { type: string }
      responses:
        '201': { description: Créé }
...
  /api/maintenance/contracts/vehicle/{vehicleId}:
    get:
      summary: Contrats d'un véhicule
      tags: [Entretien]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Liste }
...
  /api/vehicle3d/vehicle/{vehicleId}:
    get:
      summary: Modèle 3D du véhicule
      description: Retourne format et URL du fichier glTF/GLB (iOS, Android, Web)
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
        '404': { description: Modèle non disponible }
    post:
      summary: Upload modèle 3D
      description: multipart/form-data - model (fichier .glb ou .gltf)
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                model: { type: string, format: binary }
      responses:
        '201': { description: Créé }
    delete:
      summary: Supprimer le modèle 3D
      tags: [Modèle 3D]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Supprimé }